Announce features: v7, v6
