+++ run-integration-tests-v6.sh
@@ -3,4 +3,4 @@
 ZABBIX_API_URL=http://localhost:3080/api_jsonrpc.php \
 ZABBIX_API_USER=Admin \
 ZABBIX_API_PASSWORD=zabbix \
-cargo test
+cargo test --features v6
