+++ src/client/mod.rs
@@ -1,4 +1,6 @@
 pub mod client;
-pub mod jsonrpc;
 pub mod post;
+pub mod response;
+
+#[cfg(feature = "v6")]
 pub mod v6;
