+++ src/client/v6/mod.rs
@@ -1,1 +1,2 @@
 pub mod client;
+pub mod request;
