+++ src/client/v6/request.rs
@@ -1,0 +1,10 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct ZabbixApiRequest<T: Serialize> {
+    pub jsonrpc: String,
+    pub method: String,
+    pub params: T,
+    pub id: i8,
+    pub auth: Option<String>,
+}
